+++ Source/Assimil8or/Preset/ZoneProperties.h
@@ -17,34 +17,34 @@ public:
     void setIndex (int index, bool includeSelfCallback);
     void setLevelOffset (double levelOffset, bool includeSelfCallback);
     void setLoopLength (double loopLength, bool includeSelfCallback);
-    void setLoopStart (int loopStart, bool includeSelfCallback);
+    void setLoopStart (int64_t loopStart, bool includeSelfCallback);
     void setMinVoltage (double minVoltage, bool includeSelfCallback);
     void setPitchOffset (double pitchOffset, bool includeSelfCallback);
     void setSample (juce::String sampleFileName, bool includeSelfCallback);
-    void setSampleStart (int sampleStart, bool includeSelfCallback);
-    void setSampleEnd (int sampleEnd, bool includeSelfCallback);
+    void setSampleStart (int64_t sampleStart, bool includeSelfCallback);
+    void setSampleEnd (int64_t sampleEnd, bool includeSelfCallback);
     void setSide (int side, bool includeSelfCallback);
 
     int getIndex ();
     double getLevelOffset ();
     double getLoopLength ();
-    int getLoopStart ();
+    int64_t getLoopStart ();
     double getMinVoltage ();
     double getPitchOffset ();
     juce::String getSample ();
-    int getSampleStart ();
-    int getSampleEnd ();
+    int64_t getSampleStart ();
+    int64_t getSampleEnd ();
     int getSide ();
 
     std::function<void (int index)> onIndexChange;
     std::function<void (double levelOffset)> onLevelOffsetChange;
     std::function<void (double loopLength)> onLoopLengthChange;
-    std::function<void (int loopStart)> onLoopStartChange;
+    std::function<void (int64_t loopStart)> onLoopStartChange;
     std::function<void (double minVoltage)> onMinVoltageChange;
     std::function<void (double pitchOffset)> onPitchOffsetChange;
     std::function<void (juce::String sampleFileName)> onSampleChange;
-    std::function<void (int sampleStart)> onSampleStartChange;
-    std::function<void (int sampleEnd)> onSampleEndChange;
+    std::function<void (int64_t sampleStart)> onSampleStartChange;
+    std::function<void (int64_t sampleEnd)> onSampleEndChange;
     std::function<void (int side)> onSideChange;
 
     static juce::ValueTree create (int index);
